Added a Player Selection Window
This allows the players to choose whether they would want to play against an AI or another human player. As of now, the AI has not been implemented.

Additional Bug Fixes and a new 'On Closing' function that displays a message to the user when they try to close the game.